backend/requirements.txt
# Django + DRF 의존성
django>=5.1
djangorestframework>=3.15
django-environ>=0.11
dj-database-url>=2.1
django-redis>=5.4
psycopg2-binary>=2.9
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
redis>=5.0.0
google-genai>=1.0.0
groq>=0.4.0
openai>=1.0.0
gunicorn>=22.0